requirements.txt
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
python-docx>=1.1.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
openai>=1.0.0
